user-info-api/requirements.txt
Flask==3.0.0
PyYAML==6.0.1
flask-cors==4.0.0
gunicorn==21.2.0
